Small int -> long bug corrected in ReversiblePomoRateMatrixFunction.

+++ src/core/functions/phylogenetics/ratematrix/ReversiblePomoRateMatrixFunction.h
@@ -51,7 +51,7 @@ namespace RevBayesCore {
 
         // members
 
-        const TypedDagNode< int >*                          populationSize;
+        const TypedDagNode< long >*                          populationSize;
         // const TypedDagNode< RateGenerator >*                mutationMatrix;
         const TypedDagNode< RbVector<double> >*                   exchangeabilities;
         const TypedDagNode< Simplex >*                   equilibriumFrequencies;
